+++ src/main/java/org/programmers/VoucherManagement/VoucherManagementRunner.java
@@ -1,25 +1,18 @@
 package org.programmers.VoucherManagement;
 
 import lombok.RequiredArgsConstructor;
+import org.programmers.VoucherManagement.io.CommandExecutor;
 import org.programmers.VoucherManagement.io.CommandType;
 import org.programmers.VoucherManagement.io.Console;
-import org.programmers.VoucherManagement.member.api.MemberController;
-import org.programmers.VoucherManagement.member.dto.GetMemberListRes;
-import org.programmers.VoucherManagement.voucher.api.VoucherController;
-import org.programmers.VoucherManagement.voucher.domain.DiscountType;
-import org.programmers.VoucherManagement.voucher.dto.CreateVoucherReq;
-import org.programmers.VoucherManagement.voucher.dto.GetVoucherListRes;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
 
 @Component
 @RequiredArgsConstructor
 public class VoucherManagementRunner implements CommandLineRunner {
     private final Console console;
-    private final VoucherController voucherController;
-    private final MemberController memberController;
+    private final CommandExecutor commandExecutor;
 
     @Override
     public void run(String... args) throws Exception {
@@ -34,36 +27,8 @@ public class VoucherManagementRunner implements CommandLineRunner {
                 isEnd = true;
             }
 
-            execute(commandType);
+            commandExecutor.execute(commandType);
         }
-    }
-
-    private void execute(CommandType commandType) {
-        switch (commandType) {
-            case CREATE -> {
-                console.printDiscountType();
-                CreateVoucherReq request = makeCreateVoucherRequest(); // 유효성까지 다 되어있어서 request setting 되어이썽야함.
-                voucherController.createVoucher(request);
-            }
-            case LIST -> {
-               GetVoucherListRes voucherList = voucherController.getVoucherList();
-                console.printVoucherList(voucherList);
-            }
-            case EXIT -> {
-                console.printExitMessage();
-            }
-            case BLACKLIST -> {
-                GetMemberListRes blackMemberList = memberController.getBlackMemberList();
-                console.printMemberList(blackMemberList);
-            }
-        }
-    }
-
-    private CreateVoucherReq makeCreateVoucherRequest() {
-        DiscountType discountType = console.readDiscountType();
-        int discountValue = console.readDiscountValue(discountType);
-
-        return new CreateVoucherReq(discountType, discountValue);
     }
 
 }
